move to common component for event banner contents
