+++ src/components/EventBannerElement.svelte
@@ -1,0 +1,213 @@
+<script>
+    import { EventIsOngoing } from './store.js';
+    import ExportEventButton from './ExportEventButton.svelte';
+    import VideoButton from './VideoButton.svelte';
+    export let events = [];
+    export let event_time_category = '';
+
+    export let event_type_classes = {};
+    export let event_type_icons = {};
+
+    let backgroundIcon = '';
+
+    if (event_time_category === 'upcoming') {
+        backgroundIcon = 'fa-alarm-clock';
+        events = events.filter((event) => {
+            let time_window = 70 * 24 * 60 * 60 * 1000; //TODO: change back to 1 day
+            const event_start_unix = event.data.start.getTime();
+            const event_end_unix = event.data.end.getTime();
+
+            // increase time window to a week for events longer than 5 hours
+            if (event_end_unix - event_start_unix > 5 * 60 * 60 * 1000) {
+                time_window = 70 * 24 * 60 * 60 * 1000; //TODO: change back to 7 days
+            }
+            if (event.data.start < new Date() && new Date() < event.data.end) {
+                return false;
+            }
+
+            if (event_start_unix < new Date().getTime() + time_window && new Date().getTime() < event_end_unix) {
+                // TODO: uncoment above line and remove below line, which is only for testing purposes
+                // if (new Date().getTime() < event_end_unix) {
+                return true;
+            }
+        });
+    } else if (event_time_category === 'ongoing') {
+        backgroundIcon = 'fa-broadcast-tower';
+        events = events
+            .filter((event) => {
+                return event.data.start < new Date() && new Date() < event.data.end;
+            })
+            .sort((a, b) => {
+                return a.data.start - b.data.start;
+            });
+
+        if (events.length > 0) {
+            EventIsOngoing.set(true);
+        } else {
+            EventIsOngoing.set(false);
+        }
+    }
+
+    let heading_title = event_time_category.charAt(0).toUpperCase() + event_time_category.slice(1) + ' event';
+    heading_title = events.length > 1 ? heading_title + 's' : heading_title;
+
+    // countdown function to event start which updates every second, by making `now` and the function reactive
+    let now = '';
+    setInterval(() => (now = new Date().getTime()), 1000);
+    $: countdown = (event_start, short = false) => {
+        if (event_time_category !== 'upcoming') {
+            return '';
+        }
+        let timeLeftString = '';
+        const distance = event_start.getTime() - now;
+        // Time calculations for days, hours, minutes and seconds
+        let days = Math.floor(distance / (1000 * 60 * 60 * 24));
+        const hours = Math.floor((distance % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
+        const minutes = Math.floor((distance % (1000 * 60 * 60)) / (1000 * 60));
+        const seconds = Math.floor((distance % (1000 * 60)) / 1000);
+        // Display the result in the element with id="demo"
+        timeLeftString = days + ' days,</br> ' + hours + 'h ' + minutes + 'm';
+        if (short) {
+            timeLeftString = days + 'd ' + hours + 'h ' + minutes + 'm';
+        }
+        // If the count down is finished, write some text
+        if (days <= 0) {
+            timeLeftString = hours + 'h ' + minutes + 'm ' + seconds + 's';
+            if (hours <= 0) {
+                timeLeftString = minutes + 'm ' + seconds + 's';
+            }
+        }
+        if (distance < 0) {
+            timeLeftString = 'Event started';
+        }
+        return timeLeftString;
+    };
+</script>
+
+{#if events.length > 0}
+    <div class={event_time_category + '-event event-container border-bottom border-black-subtle'}>
+        <div>
+            <div class="d-none d-lg-flex">
+                <div class="col-lg-4 overflow-hidden ps-3 position-relative d-flex flex-column justify-content-center">
+                    <h4 class="display-4 p-2 flex-grow-1">{heading_title}</h4>
+                    <i
+                        class={`fad ${backgroundIcon} homepage-header-fa-background mt-5 ms-1 ms-xl-5`}
+                        aria-hidden="true"
+                    />
+                </div>
+                <div class="flex-grow-1">
+                    {#each events as event}
+                        <div class=" d-flex align-items-center justify-content-between">
+                            <div class="pt-lg-3 pb-lg-3 text-lg-start">
+                                <h5 class="pt-2 pb-0 pb-lg-1">
+                                    <a href={'events/' + event.slug} class="text-success text-decoration-none"
+                                        >{event.data.title}</a
+                                    >
+                                </h5>
+                                <p class="lead">
+                                    <a href={'events/' + event.slug} class="text-body text-decoration-none"
+                                        >{event.data.subtitle}</a
+                                    >
+                                </p>
+
+                                <p class="">
+                                    <a href={'events/' + event.slug} class="text-secondary-emphasis text-decoration-none"
+                                        >{event.data.duration}</a
+                                    ><span class="ms-1">
+                                        <span class={'badge bg-' + event_type_classes[event.data.type] + ' small'}
+                                            ><i
+                                                class={event_type_icons[event.data.type] + ' me-1'}
+                                                aria-hidden="true"
+                                            />
+                                            {event.data.type}</span
+                                        >
+                                    </span>
+                                </p>
+                                {#if event_time_category === 'upcoming'}
+                                    <div class="btn-group" role="group" aria-label="Event details">
+                                        <a href={'events/' + event.slug} class="btn btn-outline-success text-nowrap">
+                                            Event Details
+                                        </a>
+                                        <ExportEventButton frontmatter={event.data} />
+                                    </div>
+                                {/if}
+                            </div>
+
+                            <div class="mx-3">
+                                {#if event_time_category === 'upcoming'}
+                                    <div class="text-center text-nowrap">
+                                        <h5>Event starts in:</h5>
+                                        <span class="display-5">
+                                            {@html countdown(event.data.start)}
+                                        </span>
+                                    </div>
+                                {/if}
+                                {#if event_time_category === 'ongoing'}
+                                    <div class="">
+                                        <div class="btn-group text-nowrap" role="group" aria-label="Event details">
+                                            <a href={'events/' + event.slug} class="btn btn-outline-success"
+                                                >Event Details</a
+                                            >
+                                            <VideoButton urls={event.data.location_url} />
+                                        </div>
+                                    </div>
+                                {/if}
+                            </div>
+                        </div>
+                        <hr class="mx-4 my-0 py-0" />
+                    {/each}
+                </div>
+            </div>
+            <div class="d-lg-none">
+                <div class="pt-2 pb-1 mb-2 overflow-hidden mainpage-subheader-heading-header bg-body-tertiary">
+                    <h5 class="pt-2 font-weight-light text-center text-sucess">{heading_title}</h5>
+                </div>
+                {#each events as event}
+                    <div class="text-center ">
+                        <h4 class="pt-2 pb-0">
+                            <a href={'events/' + event.slug} class="text-success text-decoration-none"
+                                >{event.data.title}</a
+                            >
+                        </h4>
+                        <p class="d-sm-none mb-2">
+                            <a href={'events/' + event.slug} class="text-body text-decoration-none"
+                                >{event.data.subtitle}</a
+                            ><span class={'badge bg-' + event_type_classes[event.data.type] + ' small ms-3'}
+                                ><i class={event_type_icons[event.data.type] + ' me-1'} aria-hidden="true" />
+                                {event.data.type}</span
+                            >
+                        </p>
+                        <div class="small mb-2">
+                            <a href={'events/' + event.slug} class="text-secondary-emphasis text-decoration-none"
+                                >{event.data.duration}</a
+                            >
+                            <div class="btn-group text-nowrap" role="group" aria-label="Event details">
+                                <a href={'events/' + event.slug} class="btn btn-outline-success">
+                                    Event Details
+                                </a>
+                                {#if event_time_category === 'upcoming'}
+                                    <ExportEventButton frontmatter={event.data} />
+                                {/if}
+                                {#if event_time_category === 'ongoing'}
+                                    <VideoButton urls={event.data.location_url} />
+                                {/if}
+                            </div>
+                        </div>
+                    </div>
+                {/each}
+            </div>
+        </div>
+    </div>
+{/if}
+
+<style lang="scss">
+    .homepage-header-fa-background {
+        position: absolute;
+        font-size: 14em;
+        opacity: 0.2;
+    }
+    hr:last-child {
+        display: none;
+    }
+
+</style>
