+++ src/styles/main.scss
@@ -182,7 +182,7 @@ section:before {
 
 .triangle-down,
 .triangle-up {
-  margin-top: -1px;
+  margin-top: -2px;
   padding-bottom: 4rem;
   clip-path: polygon(100% 0%, 100% 5%, 50% 100%, 0 5%, 0% 0%);
 }
